fiks bug rundt mapping av sanityvilkår og oppdater test på brevbegrunnelser
